[fixes]  - social tab now correctly appears  - loading loadouts for deceased pawns no longer crash the scribe
[misc]
 - moved all loadout related code into a subfolder of its own
