+++ Source/CombatRealism/Combat_Realism/Comps/CompAmmoUser.cs
