+++ Source/CombatRealism/Combat_Realism/Comps/CompInventory.cs
