+++ Source/CombatRealism/Combat_Realism/Utility.cs
